+++ src/main/java/com/fastcampus/config/CorsConfig.java
@@ -1,4 +1,4 @@
-package com.fastcampus.Security;
+package com.fastcampus.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
